Fix counter overflow issues

+++ stats.h
@@ -13,7 +13,7 @@ class Stats {
     void terse(FILE *out) {
       gettimeofday(&end, NULL);
       float t = (float) tsdiff(&end, &start) / 1000.0;
-      fprintf(out, "# %d/%d (%2.1f%%), NBskip: %d/%d TBAout: %d/%d/%d Bad: %d Fill: %d",
+      fprintf(out, "# %llu/%llu (%2.1f%%), NBskip: %llu/%llu TBAout: %llu/%llu/%llu Bad: %llu Fill: %llu",
         count, to_probe, (float) count*100.0/to_probe,
         nbr_skipped, bgp_skipped, ttl_outside, 
         bgp_outside, adr_outside, baddst, fills);    
@@ -28,28 +28,28 @@ class Stats {
       char s[1000];
       strftime(s, 1000, "%a, %d %b %Y %T %z", p);
       fprintf(out, "# End: %s\n", s);
-      fprintf(out, "# Bad_Resp: %u\n", baddst);
-      fprintf(out, "# Fills: %u\n", fills);
-      fprintf(out, "# Outside_TTL: %u\n", ttl_outside);
-      fprintf(out, "# Outside_BGP: %u\n", bgp_outside);
-      fprintf(out, "# Outside_Addr: %u\n", adr_outside);
-      fprintf(out, "# Skipped_Nbr: %u\n", nbr_skipped);
-      fprintf(out, "# Skipped_BGP: %u\n", bgp_skipped);
-      fprintf(out, "# Pkts: %u\n", count);
+      fprintf(out, "# Bad_Resp: %llu\n", baddst);
+      fprintf(out, "# Fills: %llu\n", fills);
+      fprintf(out, "# Outside_TTL: %llu\n", ttl_outside);
+      fprintf(out, "# Outside_BGP: %llu\n", bgp_outside);
+      fprintf(out, "# Outside_Addr: %llu\n", adr_outside);
+      fprintf(out, "# Skipped_Nbr: %llu\n", nbr_skipped);
+      fprintf(out, "# Skipped_BGP: %llu\n", bgp_skipped);
+      fprintf(out, "# Pkts: %llu\n", count);
       fprintf(out, "# Elapsed: %2.2fs\n", t);
       fprintf(out, "# PPS: %2.2f\n", (float) count / t);
       fprintf(out, "#\n");
     };
     
-    uint32_t count;       // number of probes sent
-    uint32_t to_probe;
-    uint32_t nbr_skipped; // b/c already in learned neighborhood 
-    uint32_t bgp_skipped; // b/c BGP learned
-    uint32_t ttl_outside; // b/c outside range of TTLs we want
-    uint32_t bgp_outside; // b/c not in BGP table
-    uint32_t adr_outside; // b/c address outside range we want
-    uint32_t baddst;      // b/c checksum invalid on destination in reponse
-    uint32_t fills;       // extra tail probes past maxttl
+    uint64_t count;       // number of probes sent
+    uint64_t to_probe;
+    uint64_t nbr_skipped; // b/c already in learned neighborhood 
+    uint64_t bgp_skipped; // b/c BGP learned
+    uint64_t ttl_outside; // b/c outside range of TTLs we want
+    uint64_t bgp_outside; // b/c not in BGP table
+    uint64_t adr_outside; // b/c address outside range we want
+    uint64_t baddst;      // b/c checksum invalid on destination in reponse
+    uint64_t fills;       // extra tail probes past maxttl
    
     struct timeval start;
     struct timeval end;
